Further fixes to get ScoreFeatureSet to work.
